Separates trackedDB concerns into their own files. This makes for less clutter in the worker logic.
